orion-ld: avoid ABI incompatibility warning
We get several of these warnings when compiling with GCC >= 7.1:

$TOOLCHAIN/include/c++/10.2.1/bits/vector.tcc:426:7: note: parameter passing for argument of type 'std::vector<double>::iterator' changed in GCC 7.1
  426 |       vector<_Tp, _Alloc>::
      |       ^~~~~~~~~~~~~~~~~~~

There has been an ABI compatibility issue when compiling with GCC6 (had
an issue) and GCC7 (which has been fixed). As long as all code is
compiled with the same GCC version, we should be able to ignore the
warning by adding "-Wno-psabi".